feat: (network) increase abstraction level of the connection classes to allow for browser implementation
