+++ packages/network/src/connection/WebRtcEndpoint.ts
@@ -1,9 +1,8 @@
 import { EventEmitter } from 'events'
 import { Event, IWebRtcEndpoint } from './IWebRtcEndpoint'
-import nodeDataChannel, { DescriptionType } from 'node-datachannel'
 import { Logger } from '../helpers/Logger'
 import { PeerInfo } from './PeerInfo'
-import { Connection, ConstructorOptions, DeferredConnectionAttempt, isOffering } from './Connection'
+import { WebRtcConnection, ConstructorOptions, DeferredConnectionAttempt, isOffering } from './WebRtcConnection'
 import { Metrics, MetricsContext } from '../helpers/MetricsContext'
 import {
     AnswerOptions,
@@ -27,12 +26,12 @@ class WebRtcError extends Error {
     }
 }
 
-export class WebRtcEndpoint extends EventEmitter implements IWebRtcEndpoint {
+export abstract class WebRtcEndpoint extends EventEmitter implements IWebRtcEndpoint {
     private readonly peerInfo: PeerInfo
     private readonly stunUrls: string[]
     private readonly rtcSignaller: RtcSignaller
     private readonly negotiatedProtocolVersions: NegotiatedProtocolVersions
-    private connections: { [key: string]: Connection }
+    private connections: { [key: string]: WebRtcConnection }
     private messageQueues: { [key: string]: MessageQueue<string> }
     private readonly newConnectionTimeout: number
     private readonly pingInterval: number
@@ -129,7 +128,7 @@ export class WebRtcEndpoint extends EventEmitter implements IWebRtcEndpoint {
             pingInterval: this.pingInterval,
         }
 
-        const connection = new Connection(connectionOptions)
+        const connection = this.doCreateConnection(connectionOptions)
 
         if (connection.isOffering()) {
             connection.once('localDescription', (type, description) => {            
@@ -181,7 +180,7 @@ export class WebRtcEndpoint extends EventEmitter implements IWebRtcEndpoint {
     private onRtcOfferFromSignaller({ routerId, originatorInfo, description, connectionId }: OfferOptions): void {
         const { peerId } = originatorInfo
         
-        let connection: Connection
+        let connection: WebRtcConnection
        
         if (!this.connections[peerId]) {
             connection = this.createConnection(peerId, routerId,null)
@@ -201,7 +200,7 @@ export class WebRtcEndpoint extends EventEmitter implements IWebRtcEndpoint {
         }
         connection.setPeerInfo(PeerInfo.fromObject(originatorInfo))
         connection.setConnectionId(connectionId)
-        connection.setRemoteDescription(description, 'offer' as DescriptionType.Offer)
+        connection.setRemoteDescription(description, 'offer')
     }
 
     private onRtcAnswerFromSignaller({ originatorInfo, description, connectionId }: AnswerOptions): void {
@@ -213,7 +212,7 @@ export class WebRtcEndpoint extends EventEmitter implements IWebRtcEndpoint {
             this.logger.warn('unexpected rtcAnswer from %s (connectionId mismatch %s !== %s)', peerId, connection.getConnectionId(), connectionId)
         } else {
             connection.setPeerInfo(PeerInfo.fromObject(originatorInfo))
-            connection.setRemoteDescription(description, 'answer' as DescriptionType.Answer)
+            connection.setRemoteDescription(description, 'answer')
             this.attemptProtocolVersionValidation(connection)
         }
     }
@@ -254,7 +253,7 @@ export class WebRtcEndpoint extends EventEmitter implements IWebRtcEndpoint {
         }
     }
 
-    private replaceConnection(peerId: string, routerId: string, newConnectionId?: string): Connection {
+    private replaceConnection(peerId: string, routerId: string, newConnectionId?: string): WebRtcConnection {
         // Close old connection
         const conn = this.connections[peerId]
         let deferredConnectionAttempt = null
@@ -349,7 +348,7 @@ export class WebRtcEndpoint extends EventEmitter implements IWebRtcEndpoint {
         this.metrics.record('msgOutSpeed', 1)
     }
 
-    private attemptProtocolVersionValidation(connection: Connection): void {
+    private attemptProtocolVersionValidation(connection: WebRtcConnection): void {
         try {
             this.negotiatedProtocolVersions.negotiateProtocolVersion(
                 connection.getPeerId(),
@@ -422,6 +421,9 @@ export class WebRtcEndpoint extends EventEmitter implements IWebRtcEndpoint {
         this.removeAllListeners()
         Object.values(connections).forEach((connection) => connection.close())
         Object.values(messageQueues).forEach((queue) => queue.clear())
-        nodeDataChannel.cleanup()
+        this.doStop()
     }
+
+    protected abstract doCreateConnection(opts: ConstructorOptions): WebRtcConnection
+    protected abstract doStop(): void
 }
