+++ packages/network/test/integration/webrtc-endpoint-back-pressure-handling.test.ts
@@ -1,5 +1,5 @@
 import { Event } from '../../src/connection/IWebRtcEndpoint'
-import { WebRtcEndpoint } from '../../src/connection/WebRtcEndpoint'
+import { NodeWebRtcEndpoint } from '../../src/connection/NodeWebRtcEndpoint'
 import { PeerInfo } from '../../src/connection/PeerInfo'
 import { MetricsContext } from '../../src/helpers/MetricsContext'
 import { RtcSignaller } from '../../src/logic/RtcSignaller'
@@ -14,8 +14,8 @@ describe('WebRtcEndpoint: back pressure handling', () => {
     let tracker: Tracker
     let trackerNode1: TrackerNode
     let trackerNode2: TrackerNode
-    let ep1: WebRtcEndpoint
-    let ep2: WebRtcEndpoint
+    let ep1: NodeWebRtcEndpoint
+    let ep2: NodeWebRtcEndpoint
 
     beforeEach(async () => {
         tracker = await startTracker({
@@ -36,14 +36,14 @@ describe('WebRtcEndpoint: back pressure handling', () => {
         await trackerNode2.connectToTracker(tracker.getAddress())
 
         // Set up WebRTC endpoints
-        ep1 = new WebRtcEndpoint(
+        ep1 = new NodeWebRtcEndpoint(
             peerInfo1,
             ['stun:stun.l.google.com:19302'],
             new RtcSignaller(peerInfo1, trackerNode1),
             new MetricsContext('ep1'),
             new NegotiatedProtocolVersions(peerInfo1)
         )
-        ep2 = new WebRtcEndpoint(
+        ep2 = new NodeWebRtcEndpoint(
             peerInfo2,
             ['stun:stun.l.google.com:19302'],
             new RtcSignaller(peerInfo2, trackerNode2),
